refactor: moves time slots service to correct folder
Updates import paths to reflect the new file structure.

This change improves project organization by grouping service-related files within a dedicated "services" directory, enhancing maintainability and clarity.
